* Fixed automation.actions.ts and moved TRIGGER_LABELS const to constants * Added subscriptionPlans and paymentProviders constants
